% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezlm.R
\name{ezlm}
\alias{ezlm}
\title{Easy-to-read Linear Model}
\usage{
ezlm(dataset, iv, dv, digits, ...)
}
\arguments{
\item{dataset}{The data on which the modelling process will be run.
This parameter is so named to indicate to the user that a dataset is required}

\item{iv}{The variable to be passed to the modelling function as predictor.
This parameter is so named to avoid any confusion about directionality.}

\item{dv}{The variable to be passed to the modelling function as response.
As above, this parameter is so named to avoid confusion about directionality.}

\item{digits}{The number of digits to display for all computed numbers.
This parameter is so named to indicate to the user that a number of digits
is required}

\item{...}{Additional arguments to lower level functions}
}
\value{
Selected results of a linear regression on provided data in an
easy-to-read format composed of text and rounded digits.
}
\description{
Fits a simple linear regression model, outputting easy-to-read
results rounded to user's preferred decimal place. A wrapper for \code{lm()}.
}
\examples{
ezlm(mtcars, mtcars$mpg, mtcars$hp, 3)
ezlm(USArrests, USArrests$UrbanPop, USArrests$Assault, 2)

}
